pyproject.toml
[tool.poetry]
name = "insiderweek"
version = "0.1.0"
description = ""
authors = ["Матвей <motr.89@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.1.2"
Flask-SQLAlchemy = "^2.5.1"
Flask-Login = "^0.6.1"
bokeh = "^2.4.3"
dramatiq = {extras = ["redis"], version = "^1.13.0"}

[tool.poetry.dev-dependencies]
black = "22.3.0"
isort = "^5.10.1"
autoflake = "^1.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
